// src/interfaces/post.d.ts
export declare interface Post{
    slug?: string,
    title?: string,
    description?: string,
    publishTime?: number,
    lastUpdatedTime?: number,
    tags?: string[],
    category?: string,
    plainContent?: string,
    mdContent?: string,
    wordCount?: number,
    bannerImg?: string,
    cachedHtml?: string,
    coverFit?: string,
    context?: string,
}